[feature] #111 UniverseMarkerDrawer
